requirements.txt
ipykernel
networkx
pandas
python-gedcom